% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPAS.R
\name{imputation}
\alias{imputation}
\title{The function of imputaion.}
\usage{
imputation(obj, assay = "RNA", method = c("KNN", "ALRA"))
}
\arguments{
\item{obj}{A seurat object.}

\item{assay}{The assay for imputation. The default is 'RNA'.}

\item{method}{The method for imputation. The default is 'RNA'.}
}
\value{
A seurat object after imputaion.
}
\description{
The function of imputaion.
}
