% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12-imputation.R
\name{imputation2}
\alias{imputation2}
\title{The function of imputaion.}
\usage{
imputation2(
  obj,
  assay = "RNA",
  method = c("KNN", "ALRA"),
  verbose = SigBridgeRUtils::getFuncOption("verbose")
)
}
\arguments{
\item{obj}{A seurat object.}

\item{assay}{The assay for imputation. The default is 'RNA'.}

\item{method}{The method for imputation. The default is 'RNA'.}

\item{verbose}{Logical, whether to print messages}
}
\value{
A seurat object after imputaion.
}
\description{
The function of imputaion.
}
\seealso{
Other scPAS: 
\code{\link{imputation_ALRA2}()},
\code{\link{imputation_KNN2}()},
\code{\link{scPAS.optimized}()}

Other imputation: 
\code{\link{imputation_ALRA2}()},
\code{\link{imputation_KNN2}()}
}
\concept{imputation}
\concept{scPAS}
