% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPAS.R
\name{imputation_ALRA}
\alias{imputation_ALRA}
\title{A method for imputation of missing values in single cell RNA-sequencing data based on ALRA.}
\usage{
imputation_ALRA(obj, assay = "RNA")
}
\arguments{
\item{obj}{A seurat object.}

\item{assay}{The assay for imputation. The default is 'RNA'.}
}
\value{
A seurat object after imputaion.
}
\description{
A method for imputation of missing values in single cell RNA-sequencing data based on ALRA.
}
