% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12-imputation.R
\name{imputation_KNN2}
\alias{imputation_KNN2}
\title{A method for imputation of missing values in single cell RNA-sequencing data based on the average expression value of nearest neighbor cells.}
\usage{
imputation_KNN2(obj, assay = "RNA", LogNormalized = TRUE)
}
\arguments{
\item{obj}{A seurat object.}

\item{assay}{The assay for imputation. The default is 'RNA'.}

\item{LogNormalized}{Whether the data is LogNormalized.}
}
\value{
A seurat object after imputaion.
}
\description{
A method for imputation of missing values in single cell RNA-sequencing data based on the average expression value of nearest neighbor cells.
}
\seealso{
Other scPAS: 
\code{\link{imputation2}()},
\code{\link{imputation_ALRA2}()},
\code{\link{scPAS.optimized}()}

Other imputation: 
\code{\link{imputation2}()},
\code{\link{imputation_ALRA2}()}
}
\concept{imputation}
\concept{scPAS}
