% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-scPAS.R
\name{optimize_model}
\alias{optimize_model}
\title{Optimize Model Parameters}
\usage{
optimize_model(x, y, Network, alpha, cutoff, family, seed, verbose)
}
\description{
Optimize Model Parameters
}
\keyword{internal}
