% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-scPAS.R
\name{prepare_phenotype}
\alias{prepare_phenotype}
\title{Prepare Phenotype Data Based on Regression Family}
\usage{
prepare_phenotype(phenotype, family, tag, verbose)
}
\description{
Prepare Phenotype Data Based on Regression Family
}
\keyword{internal}
