% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPAS.R
\name{sparse.cor}
\alias{sparse.cor}
\title{A function compute the correlation of a sparse matrix.}
\usage{
sparse.cor(x)
}
\arguments{
\item{x}{Matrix. Normalized single cell expression profile extracted from Seurat object.}
}
\value{
A correlation matrix.
}
\description{
A function compute the correlation of a sparse matrix.
}
