% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-scPAS.R
\name{calculate_risk_score}
\alias{calculate_risk_score}
\title{Calculate Statistics and Cell Labels}
\usage{
calculate_risk_score(
  risk_score,
  mean.background,
  sd.background,
  FDR.threshold,
  cell_names
)
}
\description{
Calculate Statistics and Cell Labels
}
\keyword{internal}
