% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-scPAS.R
\name{identify_common_genes}
\alias{identify_common_genes}
\title{Identify Common Genes Between Datasets}
\usage{
identify_common_genes(
  bulk_dataset,
  sc_dataset,
  nfeature = NULL,
  verbose = TRUE
)
}
\description{
Identify Common Genes Between Datasets
}
\keyword{internal}
