% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPAS.R
\name{imputation_KNN}
\alias{imputation_KNN}
\title{A method for imputation of missing values in single cell RNA-sequencing data based on the average expression value of nearest neighbor cells.}
\usage{
imputation_KNN(obj, assay = "RNA", LogNormalized = T)
}
\arguments{
\item{obj}{A seurat object.}

\item{assay}{The assay for imputation. The default is 'RNA'.}

\item{LogNormalized}{Whether the data is LogNormalized.}
}
\value{
A seurat object after imputaion.
}
\description{
A method for imputation of missing values in single cell RNA-sequencing data based on the average expression value of nearest neighbor cells.
}
