% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-scPAS.R
\name{perform_permutation_test}
\alias{perform_permutation_test}
\title{Perform Permutation Test}
\usage{
perform_permutation_test(
  scaled_exp,
  Coefs,
  permutation_times,
  independent,
  FDR.threshold,
  seed
)
}
\description{
Perform Permutation Test
}
\keyword{internal}
