% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPAS.R
\name{scPAS.prediction}
\alias{scPAS.prediction}
\title{scPAS.prediction: A function that uses the scPAS model to make predictions on independent data}
\usage{
scPAS.prediction(
  model,
  test.data,
  assay = "RNA",
  FDR.threshold = 0.05,
  imputation = F,
  imputation_method = "KNN",
  independent = F
)
}
\arguments{
\item{model}{seurat object. A seurat object containing the scPAS model}

\item{test.data}{Matrix or seurat object. Single-cell RNA-seq expression matrix of related disease. Each row represents a gene and each column represents a sample. A Seurat object that contains the preprocessed data and constructed network is preferred.}

\item{assay}{Name of Assay to get.}

\item{FDR.threshold}{Numeric. FDR value threshold for identifying phenotype-associated cells.
The default is 0.05.}
}
\value{
A seurat object or data frame containing the forecast results.
}
\description{
scPAS.prediction: A function that uses the scPAS model to make predictions on independent data
}
