% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-scPAS.prediction.R
\name{scPAS.prediction}
\alias{scPAS.prediction}
\title{Single-Cell Phenotype Association Score Prediction}
\usage{
scPAS.prediction(
  model,
  test.data,
  assay = "RNA",
  FDR.threshold = 0.05,
  imputation = F,
  imputation_method = "KNN",
  independent = T
)
}
\arguments{
\item{model}{Seurat object containing the trained scPAS model}

\item{test.data}{Matrix or Seurat object containing test single-cell RNA-seq data}

\item{assay}{Name of assay to use for prediction (default: 'RNA')}

\item{FDR.threshold}{FDR threshold for identifying phenotype-associated cells (default: 0.05)}

\item{imputation}{Whether to perform data imputation (default: FALSE)}

\item{imputation_method}{Method for data imputation (default: 'KNN')}

\item{independent}{Whether test data is independent from training (default: TRUE)}
}
\value{
Seurat object with prediction results added as metadata, or data frame
containing prediction results if input was a matrix
}
\description{
Uses a trained scPAS model to make predictions on independent single-cell data.
Calculates phenotype association scores and identifies significant cells based
on false discovery rate thresholding.
}
\examples{
\dontrun{
# Predict using trained model on new data
predictions <- scPAS.prediction(
model = trained_model,
test.data = new_sc_data,
FDR.threshold = 0.05
)
}

}
