% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPAS.R
\name{scPAS}
\alias{scPAS}
\title{scPAS : A tool for identifying Phenotype-Associated cell Subpopulations from single-cell sequencing data by integrating bulk data}
\usage{
scPAS(
  bulk_dataset,
  sc_dataset,
  phenotype,
  assay = "RNA",
  tag = NULL,
  nfeature = NULL,
  imputation = F,
  imputation_method = c("KNN", "ALRA"),
  alpha = NULL,
  network_class = c("bulk", "SC"),
  independent = F,
  cutoff = 0.2,
  family = c("gaussian", "binomial", "cox"),
  permutation_times = 2000,
  FDR.threshold = 0.05
)
}
\arguments{
\item{bulk_dataset}{Matrix. Bulk expression matrix of related disease. Each row represents a gene and each column represents a sample. The input expression values are continuous, such as microarray fluorescent units in logarithmic scale, RNA-seq log-CPMs, log-RPKMs or log-TPMs.}

\item{sc_dataset}{Matrix or seurat object. Single-cell RNA-seq expression matrix of related disease. Each row represents a gene and each column represents a sample. A Seurat object that contains the preprocessed data and constructed network is preferred. Otherwise, a cell-cell similarity network is constructed based on the input matrix.Otherwise, the raw count expression matrix will be processed by using Seurat's default parameters. See run_Seurat for details.}

\item{phenotype}{Phenotype annotation of each bulk sample. It can be a continuous dependent variable,
binary group indicator vector, or clinical survival data:
  \itemize{
  \item Continuous dependent variable. Should be a quantitative vector for \code{family = gaussian}.
  \item Binary group indicator vector. Should be either a 0-1 encoded vector or a factor with two levels for \code{family = binomial}.
  \item Clinical survival data. Should be a two-column matrix with columns named 'time' and 'status'. The latter is a binary variable,
  with '1' indicating event (e.g.recurrence of cancer or death), and '0' indicating right censored.
  The function \code{Surv()} in package survival produces such a matrix.
  }}

\item{assay}{Name of Assay to get.}

\item{tag}{Names for each phenotypic group. Used for logistic regressions only.}

\item{nfeature}{Numeric. The Number of features to select as top variable features in sc_dataset. Top variable features will be used to intersect with the features of bulk_dataset. Default is NULL.All features will be used.}

\item{imputation}{Logical. imputation or not.}

\item{imputation_method}{Character. Name of alternative method for imputation.}

\item{alpha}{Numeric. Parameter used to balance the effect of the l1 norm and the network-based penalties. It can be a number or a searching vector.
If \code{alpha = NULL}, a default searching vector is used. The range of alpha is in \code{[0,1]}. A larger alpha lays more emphasis on the l1 norm.}

\item{network_class}{The source of feature-feature similarity network. By default this is set to \code{bulk} and the other one is \code{SC}. The \code{SC} is recommended when the number of bulk data samples is small.}

\item{independent}{Logical. The background distribution of risk scores is constructed independently of each cell.}

\item{cutoff}{Numeric. Cutoff for the percentage of the scPheno selected features in total features. This parameter is used to restrict the number of the
scPheno selected features. A cutoff less than \code{50\%} (default \code{20\%}) is recommended depending on the input data.}

\item{family}{Character. Response type for the regression model.  It depends on the type of the given phenotype and
can be \code{family = gaussian} for linear regression, \code{family = binomial} for classification, or \code{family = cox} for Cox regression.}

\item{FDR.threshold}{Numeric. FDR value threshold for identifying phenotype-associated cells.
The default is 0.05.}
}
\value{
This function returns a list with the following components:
  \item{para}{ A list contains the final model parameters.}
  \item{risk_score}{ A data frame containing raw scores, scaled scores and cell classification labels.}
}
\description{
scPAS : A tool for identifying Phenotype-Associated cell Subpopulations from single-cell sequencing data by integrating bulk data
}
